fix windows build

+++ src/apps/snacman/snacman/simulations/snacgame/system/Explosion.cpp
@@ -10,7 +10,7 @@ void Explosion::update(const snac::Time & aTime)
 {
     mExplosions.each([&aTime](EntHandle aHandle, component::Explosion & aExplosion, component::Geometry & aGeometry)
     {
-        float duration = snac::asSeconds(aTime.mTimepoint - aExplosion.mStartTime);
+        float duration = (float)snac::asSeconds(aTime.mTimepoint - aExplosion.mStartTime);
 
         if (aExplosion.mParameter.isCompleted(duration))
         {
